физика, несколько новых компонентов, генерация мстности

+++ Renderer/ShaderProgram.hpp
@@ -29,7 +29,7 @@ namespace Renderer {
 		void setBool(const std::string& name, bool value) {
 			glUniform1i(glGetUniformLocation(m_ID, name.c_str()), value);
 		}
-		void setFloat(const std::string& name, float value)
+		void setFloat(const std::string& name, GLint value)
 		{
 			glUniform1i(glGetUniformLocation(m_ID, name.c_str()), value);
 		}
